+++ dca/tl_calendar_events.php
@@ -43,3 +43,11 @@ $GLOBALS['TL_DCA']['tl_calendar_events']['fields']['subscription_maximum'] = [
     'eval'      => ['rgxp' => 'digit', 'tl_class' => 'w50'],
     'sql'       => "smallint(5) unsigned NOT NULL default '0'",
 ];
+
+$GLOBALS['TL_DCA']['tl_calendar_events']['fields']['subscription_lastDay'] = [
+    'label'     => &$GLOBALS['TL_LANG']['tl_calendar_events']['subscription_lastDay'],
+    'exclude'   => true,
+    'inputType' => 'text',
+    'eval'      => ['datepicker' => true, 'rgxp' => 'datim', 'tl_class' => 'w50 wizard'],
+    'sql'       => "varchar(10) NOT NULL default ''",
+];