+++ languages/en/tl_calendar_events.php
@@ -18,6 +18,10 @@ $GLOBALS['TL_LANG']['tl_calendar_events']['subscription_maximum'] = [
     'Maximum number of subscriptions',
     'Here you can set the maximum number of subscriptions. Leave 0 to allow unlimited subscriptions.',
 ];
+$GLOBALS['TL_LANG']['tl_calendar_events']['subscription_lastDay'] = [
+    'Last day of subscription',
+    'Here you can set the last allowed day to subscribe. Leave empty to not limit the subscription time.',
+];
 
 /**
  * Legends