+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version 2.0.0 stable (????-??-??)
 - Added support for notification_center
 - Display the summary of subscriptions in the list view
 - Allow to set the maximum number of subscriptions per event
+- Allow to set the last day of subscription per event
 
 
 Version 1.0.6 stable (2015-05-26)